+++ src/StefanoTree/Exception/RootNodeAlreadyExistException.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace StefanoTree\Exception;
-
-class RootNodeAlreadyExistException extends \Exception implements ExceptionInterface
-{
-}