+++ src/StefanoTree/Exception/ValidationException.php
@@ -4,6 +4,6 @@ declare(strict_types=1);
 
 namespace StefanoTree\Exception;
 
-class ValidationException extends \Exception implements ExceptionInterface
+class ValidationException extends \RuntimeException implements ExceptionInterface
 {
 }