+++ src/StefanoTree/NestedSet/MoveStrategy/MoveStrategyInterface.php
@@ -4,13 +4,15 @@ declare(strict_types=1);
 
 namespace StefanoTree\NestedSet\MoveStrategy;
 
+use StefanoTree\Exception\ValidationException;
+
 interface MoveStrategyInterface
 {
     /**
      * @param int|string|null $sourceNodeId
      * @param int|string|null $targetNodeId
      *
-     * @return bool
+     * @throws ValidationException if was not moved
      */
-    public function move($sourceNodeId, $targetNodeId): bool;
+    public function move($sourceNodeId, $targetNodeId): void;
 }