+++ src/StefanoTree/NestedSet/Validator/ValidatorInterface.php
@@ -4,12 +4,16 @@ declare(strict_types=1);
 
 namespace StefanoTree\NestedSet\Validator;
 
+use StefanoTree\Exception\ValidationException;
+
 interface ValidatorInterface
 {
     /**
      * Check if tree indexes, levels is not corrupted.
      *
      * @param int|string $rootNodeId
+     *
+     * @throws ValidationException if cannot validate tree
      *
      * @return bool
      */
@@ -19,6 +23,8 @@ interface ValidatorInterface
      * Rebuild broken tree left indexes, right indexes, levels.
      *
      * @param int|string $rootNodeId
+     *
+     * @throws ValidationException if cannot rebuild tree
      */
     public function rebuild($rootNodeId): void;
 }